:rocket: code: workspaces/client/src/pages/launches.tsx
Now that we have fetchMore, let's connect it to a Load More button to fetch more items when it's
clicked. To do this, we will need to specify an updateQuery function on the return object from
fetchMore that tells the Apollo cache how to update our query with the new items we're fetching.